field and headland highlights to floats

+++ SourceCode/Twol/Classes/CBoundary.cs
@@ -52,20 +52,20 @@ namespace Twol
 
             GL.BindBuffer(BufferTarget.ArrayBuffer, vbo_HeadTriangles[bndNum]);
 
-            double[] triangleVertexData = new double[bndList[bndNum].hdLineTriangleList.Count * 6];
+            float[] triangleVertexData = new float[bndList[bndNum].hdLineTriangleList.Count * 6];
 
             for (int i = 0; i < bndList[bndNum].hdLineTriangleList.Count; i++)
             {
                 // Assuming Triangle has properties or fields: A, B, C of type vec3 with .x, .y, .z
-                triangleVertexData[i * 6 + 0] = bndList[0].hdLineTriangleList[i].polygonPts[0].easting;
-                triangleVertexData[i * 6 + 1] = bndList[0].hdLineTriangleList[i].polygonPts[0].northing;
-                triangleVertexData[i * 6 + 2] = bndList[0].hdLineTriangleList[i].polygonPts[1].easting;
-                triangleVertexData[i * 6 + 3] = bndList[0].hdLineTriangleList[i].polygonPts[1].northing;
-                triangleVertexData[i * 6 + 4] = bndList[0].hdLineTriangleList[i].polygonPts[2].easting;
-                triangleVertexData[i * 6 + 5] = bndList[0].hdLineTriangleList[i].polygonPts[2].northing;
+                triangleVertexData[i * 6 + 0] = (float)bndList[0].hdLineTriangleList[i].polygonPts[0].easting;
+                triangleVertexData[i * 6 + 1] = (float)bndList[0].hdLineTriangleList[i].polygonPts[0].northing;
+                triangleVertexData[i * 6 + 2] = (float)bndList[0].hdLineTriangleList[i].polygonPts[1].easting;
+                triangleVertexData[i * 6 + 3] = (float)bndList[0].hdLineTriangleList[i].polygonPts[1].northing;
+                triangleVertexData[i * 6 + 4] = (float)bndList[0].hdLineTriangleList[i].polygonPts[2].easting;
+                triangleVertexData[i * 6 + 5] = (float)bndList[0].hdLineTriangleList[i].polygonPts[2].northing;
             }
 
-            GL.BufferData(BufferTarget.ArrayBuffer, triangleVertexData.Length * sizeof(double), triangleVertexData, BufferUsageHint.StaticDraw);
+            GL.BufferData(BufferTarget.ArrayBuffer, triangleVertexData.Length * sizeof(float), triangleVertexData, BufferUsageHint.StaticDraw);
         }
 
         public void DeleteHeadLineVertexArray(int bndNum)
